+++ src/components/UserProfile.tsx
@@ -1,28 +1,30 @@
 import React, { useEffect, useState } from 'react'
-import { UserResponse } from './usersResponse'
-import { Link, RouteComponentProps } from 'react-router-dom'
+import { UserResponse } from '../types/UsersResponse'
+import { ComponentState } from '../types/ComponentState'
+import { Link } from 'react-router-dom'
+import { History } from 'history'
 
-type ComponentState = 'loading' | 'loaded' | 'errored'
-
-interface MatchUserProfileParams {
+interface Props {
   index: string
+  history: History
 }
-export const UserProfile: React.FC<RouteComponentProps<MatchUserProfileParams>> = (props) => {
-  const [componentState, setComponentState] = useState<ComponentState>('loading')
+
+export const UserProfile: React.FC<Props> = ({ index, history }) => {
+  const [componentState, setComponentState] = useState<ComponentState>(ComponentState.LOADING)
   const [errorMessage, setErrorMessage] = useState<string>('')
   const [user, setUser] = useState<UserResponse>()
 
   function fetchUser() {
-    fetch(`/api/user-list/${props.match.params.index}`)
+    fetch(`/api/user-list/${index}`)
       .then((response) =>
         response.json().then((json: UserResponse) => {
           setUser(json)
-          setComponentState('loaded')
+          setComponentState(ComponentState.LOADED)
         }),
       )
       .catch((error) => {
         setErrorMessage(`Error fetching data from ther server: ${error}`)
-        setComponentState('errored')
+        setComponentState(ComponentState.ERROR)
       })
   }
 
@@ -30,9 +32,9 @@ export const UserProfile: React.FC<RouteComponentProps<MatchUserProfileParams>> 
     fetchUser()
   }, [])
 
-  if (componentState === 'loading') return <p>Loading...</p>
+  if (componentState === ComponentState.LOADING) return <p>Loading...</p>
 
-  if (componentState === 'errored') return <p>{errorMessage}</p>
+  if (componentState === ComponentState.ERROR) return <p>{errorMessage}</p>
 
   if (!user) return <p>Could not load user</p>
 
@@ -42,8 +44,9 @@ export const UserProfile: React.FC<RouteComponentProps<MatchUserProfileParams>> 
       <Link to="/">
         <input type="button" value="Home" />
       </Link>
-      {/* eslint-disable-next-line @typescript-eslint/no-unused-vars */}
-      <input type="button" value="Back" onClick={(__event) => props.history.go(-1)} />
+      <button value="Back" onClick={() => history.back()}>
+        Back
+      </button>
       <img src={user.picture.medium} alt="user-profile-picture" />
       <p>
         Title: {user.name.title} <input type="button" value="Edit" />
