+++ src/types/ComponentState.ts
@@ -1,0 +1,5 @@
+export enum ComponentState {
+  LOADING = 'loading',
+  LOADED = 'loaded',
+  ERROR = 'error',
+}
